Subject: SWARM-564 Update example
Motivation
———-
Currently its not possible to easily modify this example to use the underlying connection class as the driver jar is in a module.

Modifications
———-
Add the module that has been loaded for a specific driver to the deployment so its on the classpath

Result
———-
No CNFEs when modifying the example to access the underlying connection class.