SWARM-698: Example README's don't match actual code
Motivation
----------
Code inserted in the README's of the examples is no longer current with the actual code which causes problems with users trying to see what is needed to make something work.

Modifications
-------------
Updated all out of date README's and where possible converted to Asciidoc so that they could include snippets of actual code from the project to remove duplication.

Result
------
No change to behavior, its all documentation