SWARM-851: add a test that a custom main method in a WAR project is actually executed
Motivation
----------
The `jaxrs-war-main` example demonstrates that a WAR project can have
a custom `main` method, but it isn't actually tested. The existing
tests in the example wouldn't reveal a situation when the custom
`main` method would be skipped.

Modifications
-------------
The custom `main` method in the `jaxrs-war-main` example now prints
a message to stdout. The test asserts that the message is present
in the log file.

Result
------
More test coverage.