SWARM-855: add examples for integrating CDI with other Java EE technologies
Motivation
----------
CDI integrates with a bunch of other Java EE technologies.
WFLY-6941 specifically mentions:

- EJB
- JPA
- Bean Validation
- Web Services
- Transactions

Integration with JPA and transactions is already demonstrated
in examples `jpa-jaxrs-cdi/jpa-jaxrs-cdi-*`
and `jpa-jaxrs-cdi/jpa-jaxrs-cdi-jta-*`.

There should be new examples to demonstrate integration with EJB
and Bean Validation. For Web Services, it should be enough to extend
the `jaxws` example a little, because the integration space
is fairly small.

Modifications
-------------
Added examples `jaxrs/jaxrs-cdi-beanvalidation` and `jaxrs/jaxrs-cdi-ejb`.
Modified the `jaxws` example a bit.

Result
------
The examples now demonstrate all CDI integrations. This means that they
also provide sufficient test coverage for the future, when Swarm wants
to use WFLY-6941.